Move ownership of Mumble::Configuration out of Mumble::Client
Let the caller initialize its own Mumble::Configuration struct. Only
take a const reference to it in the constructor (Dependency
injection).

Since the bitrate member is modified on connection, move it to its own member
variable. That way it is also possible to access the original bitrate
from the config struct.