Refactor Mumble::CertManager
Do not write a public_key.pem anymore as it is unneccessary. The
private_key.pem contains the public key anyway and there really is no
practical use for it.

Move all the paths out to a CertificatePaths struct.

Simplify the public interface of CertManager to a single static
method.

Remove CertManager from Client, let the application pass a
const Certificate& to the Client constructor. The Client class should
not be concerned with the paths at all.

Pluginbot::Main is no responsible for creating/loading the certificate
through CertManager.