Fix Mpd::Status updating_db () bug
The libmpdclient 2.11 documentation for mpd_stats_get_update_id () is
incorrect:

The function does not only return 0 or 1. It returns 0 when
there is no database update running, but otherwise it returns the
update_id (hence the name of the function) of the currently running
database update (which is only 1 for the first update!).

Also in ruby-mpd the function is called `updating_db ()` or similar.

This lead to the false conclusion that it should actually return a
boolean.

Rename `updating_db ()` to `update_id ()` (which is consistent with
the mpd naming) and return a `uint` instead of a `bool`.

Also remove the comparison!