Improve debug text message logging
Shorten "text message: " to "> User: " where "User" is the actual user
name that send the message.

Only write a text message to log if there was an actor at all.

Do not write "no control string in text message" to log any more.

Shorten "image text" messages to 50 chars to not pollute the debug log
with BASE64 encoded images.