Fix libuv testing
Clarify the AS_CASE part of the test. That way the comments are
superfluous.

Make sure after the AS_CASE part HAVE_LIBUV is set to either 0 or 1.

Set the AC_DEFINE USE_LIBUV and AM_CONDITIONAL USE_LIBUV according to
the value of HAVE_LIBUV only. Avoid `-a` and other bash extensions.